# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api serviconsa",
        "description": "<b>Estados:</b> (Enviada, Recibida, Procesando, Finalizada, Rechazada)",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/invoice_quote": {
            "get": {
                "tags": [
                    "InvoiceQuote"
                ],
                "summary": "All invoice_quote",
                "description": "All invoice_quote",
                "operationId": "all_invoice_quote",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Process"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Cliente] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "InvoiceQuote"
                ],
                "summary": "Store ainvoice_quote",
                "description": "Store ainvoice_quote",
                "operationId": "store_invoice_quote",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "attended"
                                ],
                                "properties": {
                                    "client": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "client"
                                    },
                                    "attended": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "attended"
                                    },
                                    "taxes": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "taxes"
                                    },
                                    "discount": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "0"
                                    },
                                    "total": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "100"
                                    },
                                    "observation": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "observation"
                                    },
                                    "terms": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "terms"
                                    },
                                    "type": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "quote"
                                    },
                                    "items": {
                                        "type": "string",
                                        "format": "array",
                                        "example": [
                                            {
                                                "name": "name",
                                                "description": "description",
                                                "price": 100,
                                                "amount": 1,
                                                "total": 100
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/invoice_quote/{id}": {
            "get": {
                "tags": [
                    "InvoiceQuote"
                ],
                "summary": "See ainvoice_quote",
                "description": "See ainvoice_quote",
                "operationId": "watch_invoice_quote",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Process"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Cliente] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "InvoiceQuote"
                ],
                "summary": "Update ainvoice_quote",
                "description": "Update ainvoice_quote",
                "operationId": "update_invoice_quote",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "attended"
                                ],
                                "properties": {
                                    "client": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "client"
                                    },
                                    "attended": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "attended"
                                    },
                                    "taxes": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "taxes"
                                    },
                                    "discount": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "0"
                                    },
                                    "total": {
                                        "type": "number",
                                        "format": "number",
                                        "example": "100"
                                    },
                                    "observation": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "observation"
                                    },
                                    "terms": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "terms"
                                    },
                                    "type": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "quote"
                                    },
                                    "status": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "status"
                                    },
                                    "items": {
                                        "type": "string",
                                        "format": "array",
                                        "example": [
                                            {
                                                "name": "name",
                                                "description": "description",
                                                "price": 100,
                                                "amount": 1,
                                                "total": 100
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "InvoiceQuote"
                ],
                "summary": "Delete ainvoice_quote",
                "description": "Delete ainvoice_quote",
                "operationId": "delete_invoice_quote",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}